requirements.txt
python-slugify==8.0.1
hdx-python-api==6.0.9
-r docker-requirements.txt
